requirements.txt
numpy
tensorflow
Flask
pillow
Werkzeug
gunicorn
opencv-python
Python-IO